UpdateDiaryFragment

+++ app/src/main/java/ilya/chistousov/countcalories/data/entity/FoodDbEntity.kt
@@ -9,7 +9,7 @@ data class FoodDbEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Int,
     val name: String,
-    val calories: Double,
+    val calories: Int,
     val proteins: Double,
     val fats: Double,
     val carbs: Double,
